add logger in auth service

+++ auth/internal/api/user/delete.go
@@ -2,20 +2,23 @@ package user
 
 import (
 	"context"
-	"log"
 
 	descUser "github.com/GolZrd/micro-chat/auth/pkg/user_v1"
 
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/emptypb"
 )
 
 func (s *Implementation) Delete(ctx context.Context, req *descUser.DeleteRequest) (*emptypb.Empty, error) {
-	err := s.userService.Delete(ctx, req.Id)
-	if err != nil {
-		return nil, err
+	if req.Id <= 0 {
+		return nil, status.Error(codes.InvalidArgument, "id is required")
 	}
 
-	log.Printf("Delete user with id: %d", req.Id)
+	err := s.userService.Delete(ctx, req.Id)
+	if err != nil {
+		return nil, status.Error(codes.Internal, err.Error())
+	}
 
 	return &emptypb.Empty{}, nil
 }
